Fix JSON encoder check
When maxdepth is absent from the configuration, the
`json_instanciate` method uses the default value of
`JSON_TOKENER_DEFAULT_DEPTH` (a correct value) but `json_check` does
not initialize its variable properly and uses 0 (an incorrect value).

The fix is simply to initialize the same way in `json_check` as it is
done in `json_instanciate`.